+++ src/neoxp/Commands/CreateCommand.cs
@@ -27,7 +27,7 @@ namespace NeoExpress.Commands
         [Option(Description = "Overwrite existing data")]
         internal bool Force { get; set; }
 
-        internal int OnExecute(IConsole console)
+        internal int OnExecute(CommandLineApplication app, IConsole console)
         {
             try
             {
@@ -42,7 +42,7 @@ namespace NeoExpress.Commands
             }
             catch (Exception ex)
             {
-                console.Error.WriteLine(ex.Message);
+                app.WriteException(ex);
                 return 1;
             }
         }